Added the new batch interface. (This is meant for 1 MPI task.)

+++ SRC/single/ps3dcomm.c
@@ -732,9 +732,9 @@ int_t sscatter3dLPanels(int_t nsupers,
 				int_t krow = PROW( jb, grid );
 				if(myrow==krow){   /* diagonal block */
 					int_t nsupc = SuperSize(jb);
-					if (!(Linv_bc_ptr[ljb] = (double*)SUPERLU_MALLOC(nsupc*nsupc * sizeof(double))))
+					if (!(Linv_bc_ptr[ljb] = (float*)SUPERLU_MALLOC(nsupc*nsupc * sizeof(float))))
 					ABORT("Malloc fails for Linv_bc_ptr[ljb][]");
-					if (!(Uinv_bc_ptr[ljb] = (double*)SUPERLU_MALLOC(nsupc*nsupc * sizeof(double))))
+					if (!(Uinv_bc_ptr[ljb] = (float*)SUPERLU_MALLOC(nsupc*nsupc * sizeof(float))))
 					ABORT("Malloc fails for Uinv_bc_ptr[ljb][]");
 				}else{
 					Linv_bc_ptr[ljb] = NULL;
